+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- Sort loaded logs alphabetically
+- Group loaded logs after type
+- Sort loaded logs alphabetically and after start time
 - Show loaded logs start datetime as `YYYY-MM-DD HH:MM:SS`
 
 ## [1.23.0]